+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 #### Change Log:
+* [2025.2.6] - Added `read_list` function
+* [2025.2.6] - Added automatic overwrite for logs with same name via `logger` and set default level to `logging.INFO`
 * [2025.1.23] - Added `bin/archive-subdirectories.py` and scripts section in `setup.py`
 * [2025.1.23] - Added `gzip_file` and `archive_subdirectories` functions
 * [2024.11.19] - Added `get_executable_in_path` and `add_executables_to_environment` functions